net/mlx5: E-Switch, Unload the representors in the correct order
This fixes a bug in which changing mode to legacy while handling VXLAN
traffic causes a NULL derefrence.

When switching from switchdev to legacy all the port devices (nic and
reps) are cleaned up. Part of this cleaning process is removing the
neigh entry and the hash table containing them. however, a representor
neigh entry might be linked to the nic hash table and if the nic is
cleaned first the cleaning of the representor will end up in null deref.
Fix that by unloading the representors in the opposite order of load.

Issue: 1063609
Change-Id: I609697cc684574413ad23d53cae75ea28cef9dc5
Fixes: cb67b832921c ("net/mlx5e: Introduce SRIOV VF representors")
Signed-off-by: Shahar Klein <shahark@mellanox.com>
Reviewed-by: Roi Dayan <roid@mellanox.com>